backend/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.6.0
django-colorfield==0.11.0
pillow==11.1.0
ruff==0.9.0
sqlparse==0.5.3
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0